requirements.txt
Flask==2.0.2
line-bot-sdk==1.18.0
requests==2.26.0
pandas==1.5.3
mplfinance==0.12.7a17
gunicorn==20.1.0
numpy==1.23.5
